IINTERLOK-3154 Make parent class abstract
* Make parent CSV aggregator abstract
* Update .lgtm.yml file
* Switch from tabs to spaces :-(
