+++ .lgtm.yml
@@ -3,6 +3,8 @@ extraction:
     index:
       gradle:
         version: 5.6.3
+      build_command:
+        - ./gradlew --no-daemon -S lgtmCompile
 
 path_classifiers:
   docs:
@@ -17,3 +19,5 @@ path_classifiers:
     - ".dependabot"
     - ".github"
     - ".lgtm.yml"
+  generated:
+    exclude: "**/*.java"
