+++ src/test/java/com/adaptris/csv/aggregator/CsvAggregatorTest.java
